// apps/se/src/app/application/tramites/120603/constants/tipoPersona.enum.ts
/**
 * Enum que indica los tipos de persona de un usuario
 */
export enum TipoPersona {
    FISICA = 'persona fisica',
    MORAL = 'perosona moral',
    NO_CONTRIBUYENTE= 'No Contribuyente'
}

/**
 * Enum que representa la nacionalidad mexicana.
 */
export enum NacionalidadMexicana {
    /** Representa la opción "Sí" para nacionalidad mexicana */
    SI = 'Sí',

    /** Representa la opción "No" para nacionalidad mexicana */
    NO = 'No'
}